Dont remove profile-dummy after running unit tests

+++ backend/src/routes/recipe.routes.test.ts
@@ -197,7 +197,8 @@ const removeDummyFiles = () => {
     
         files.forEach(file => {
             const fileDir = path.join(imagePath, file);
-            if (file !== 'test_burger.jpeg' && file !== 'test_falafel.jpeg' && file !== 'test_spaghetti.jpeg') {
+            if (file !== 'test_burger.jpeg' && file !== 'test_falafel.jpeg' && file !== 'test_spaghetti.jpeg'
+                && file !== 'profile-dummy.jpg') {
                 fs.unlinkSync(fileDir);
             }
         });
